# optimize allocations on pass of offsets to pipeline