--- a/pipeline/antispam/antispammer_test.go
+++ b/pipeline/antispam/antispammer_test.go
@@ -36,7 +36,7 @@ func TestAntispam(t *testing.T) {
 	startTime := time.Now()
 	checkSpam := func(i int) bool {
 		eventTime := startTime.Add(time.Duration(i) * maintenanceInterval / 2)
-		return antispamer.IsSpam(1, "test", false, []byte(`{}`), eventTime)
+		return antispamer.IsSpam("1", "test", false, []byte(`{}`), eventTime)
 	}
 
 	for i := 1; i < threshold; i++ {
@@ -66,7 +66,7 @@ func TestAntispamAfterRestart(t *testing.T) {
 	startTime := time.Now()
 	checkSpam := func(i int) bool {
 		eventTime := startTime.Add(time.Duration(i) * maintenanceInterval)
-		return antispamer.IsSpam(1, "test", false, []byte(`{}`), eventTime)
+		return antispamer.IsSpam("1", "test", false, []byte(`{}`), eventTime)
 	}
 
 	for i := 1; i < threshold; i++ {
@@ -128,7 +128,7 @@ func TestAntispamExceptions(t *testing.T) {
 	antispamer.exceptions.Prepare()
 
 	checkSpam := func(source, event string, wantMetric map[string]float64) {
-		antispamer.IsSpam(1, source, true, []byte(event), now)
+		antispamer.IsSpam("1", source, true, []byte(event), now)
 		for k, v := range wantMetric {
 			r.Equal(v, testutil.ToFloat64(antispamer.exceptionMetric.WithLabelValues(k)))
 		}