--- a/plugin/action/add_file_name/add_file_name_test.go
+++ b/plugin/action/add_file_name/add_file_name_test.go
@@ -21,8 +21,8 @@ func TestModify(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, sourceName, test.Offset(0), []byte(`{"error":"info about error"}`))
-	input.In(0, sourceName, test.Offset(0), []byte(`{"file":"not_my_file"}`))
+	input.In(0, sourceName, test.NewOffset(0), []byte(`{"error":"info about error"}`))
+	input.In(0, sourceName, test.NewOffset(0), []byte(`{"file":"not_my_file"}`))
 
 	wg.Wait()
 	p.Stop()