--- a/plugin/action/add_host/add_host_test.go
+++ b/plugin/action/add_host/add_host_test.go
@@ -22,7 +22,7 @@ func TestModify(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{}`))
 
 	wg.Wait()
 	p.Stop()