--- a/plugin/action/convert_log_level/convert_log_level_test.go
+++ b/plugin/action/convert_log_level/convert_log_level_test.go
@@ -235,7 +235,7 @@ func TestDo(t *testing.T) {
 			})
 
 			for _, log := range tc.In {
-				input.In(0, "test.log", test.Offset(0), []byte(log))
+				input.In(0, "test.log", test.NewOffset(0), []byte(log))
 			}
 
 			now := time.Now()