--- a/plugin/action/decode/decode_test.go
+++ b/plugin/action/decode/decode_test.go
@@ -275,7 +275,7 @@ func TestDecode(t *testing.T) {
 				wg.Done()
 			})
 
-			input.In(0, "test.log", test.Offset(0), tt.input)
+			input.In(0, "test.log", test.NewOffset(0), tt.input)
 
 			wg.Wait()
 			p.Stop()