--- a/plugin/action/flatten/flatten_test.go
+++ b/plugin/action/flatten/flatten_test.go
@@ -25,7 +25,7 @@ func TestFlatten(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"complex":{"a":"b","c":"d"}}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"complex":{"a":"b","c":"d"}}`))
 
 	wg.Wait()
 	p.Stop()