--- a/plugin/action/join/join_test.go
+++ b/plugin/action/join/join_test.go
@@ -172,7 +172,7 @@ func TestSimpleJoin(t *testing.T) {
 
 			for i := 0; i < tt.iterations; i++ {
 				for m, line := range lines {
-					input.In(0, "test.log", test.Offset(int64(i*10000+m)), []byte(line))
+					input.In(0, "test.log", test.NewOffset(int64(i*10000+m)), []byte(line))
 				}
 			}
 
@@ -271,7 +271,7 @@ func TestJoinAfterNilNode(t *testing.T) {
 
 			for i := 0; i < tt.iterations; i++ {
 				for m, line := range lines {
-					input.In(0, "test.log", test.Offset(int64(i*10000+m)), []byte(line))
+					input.In(0, "test.log", test.NewOffset(int64(i*10000+m)), []byte(line))
 				}
 			}