--- a/plugin/action/json_decode/json_decode_test.go
+++ b/plugin/action/json_decode/json_decode_test.go
@@ -19,7 +19,7 @@ func TestDecode(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"log":"{\"field2\":\"value2\",\"field3\":\"value3\"}"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"log":"{\"field2\":\"value2\",\"field3\":\"value3\"}"}`))
 
 	wg.Wait()
 	p.Stop()