--- a/plugin/action/json_encode/json_encode_test.go
+++ b/plugin/action/json_encode/json_encode_test.go
@@ -20,7 +20,7 @@ func TestEncode(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"server":{"os":"linux","arch":"amd64"}}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"server":{"os":"linux","arch":"amd64"}}`))
 
 	wg.Wait()
 	p.Stop()