--- a/plugin/action/json_extract/json_extract_test.go
+++ b/plugin/action/json_extract/json_extract_test.go
@@ -118,7 +118,7 @@ func TestJsonExtract(t *testing.T) {
 				wg.Done()
 			})
 
-			input.In(0, "test.log", test.Offset(0), []byte(tt.in))
+			input.In(0, "test.log", test.NewOffset(0), []byte(tt.in))
 
 			wg.Wait()
 			p.Stop()