--- a/plugin/action/keep_fields/keep_fields_test.go
+++ b/plugin/action/keep_fields/keep_fields_test.go
@@ -21,9 +21,9 @@ func TestKeepFields(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_1":"value_1","a":"b"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_2":"value_2","b":"c"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_3":"value_3","a":"b"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_1":"value_1","a":"b"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_2":"value_2","b":"c"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_3":"value_3","a":"b"}`))
 
 	wg.Wait()
 	p.Stop()