--- a/plugin/action/move/move_test.go
+++ b/plugin/action/move/move_test.go
@@ -256,7 +256,7 @@ func TestMove(t *testing.T) {
 				wg.Done()
 			})
 
-			input.In(0, "test.log", test.Offset(0), []byte(tt.in))
+			input.In(0, "test.log", test.NewOffset(0), []byte(tt.in))
 
 			wg.Wait()
 			p.Stop()