--- a/plugin/action/parse_es/pipeline_test.go
+++ b/plugin/action/parse_es/pipeline_test.go
@@ -110,7 +110,7 @@ func TestPipeline(t *testing.T) {
 			})
 
 			for _, event := range tCase.eventsIn {
-				input.In(event.sourceID, event.sourceName, test.Offset(event.offset), event.bytes)
+				input.In(event.sourceID, event.sourceName, test.NewOffset(event.offset), event.bytes)
 			}
 
 			wg.Wait()