--- a/plugin/action/remove_fields/remove_fields_test.go
+++ b/plugin/action/remove_fields/remove_fields_test.go
@@ -22,9 +22,9 @@ func TestRemoveFields(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_1":"value_1","a":"b"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_2":"value_2","b":"c"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_3":"value_3","a":"b"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_1":"value_1","a":"b"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_2":"value_2","b":"c"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_3":"value_3","a":"b"}`))
 
 	wg.Wait()
 	p.Stop()
@@ -47,9 +47,9 @@ func TestRemoveNestedFields(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"a":"some"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"a":{"b":"deleted"}}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"a":{"b":{"c":["deleted"]},"d":"saved"}}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"a":"some"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"a":{"b":"deleted"}}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"a":{"b":{"c":["deleted"]},"d":"saved"}}`))
 
 	wg.Wait()
 	p.Stop()