--- a/plugin/action/rename/rename_test.go
+++ b/plugin/action/rename/rename_test.go
@@ -31,11 +31,11 @@ func TestRename(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_1":"value_1"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_2":"value_2"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_3":"value_3"}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"field_4":{"field_5":"value_5"}}`))
-	input.In(0, "test.log", test.Offset(0), []byte(`{"k8s_node_label_topology.kubernetes.io/zone":"value_6"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_1":"value_1"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_2":"value_2"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_3":"value_3"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"field_4":{"field_5":"value_5"}}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"k8s_node_label_topology.kubernetes.io/zone":"value_6"}`))
 
 	wg.Wait()
 	p.Stop()
@@ -71,7 +71,7 @@ func TestRenamingSequence(t *testing.T) {
 		wg.Done()
 	})
 
-	input.In(0, "test.log", test.Offset(0), []byte(`{"key1":"value_1"}`))
+	input.In(0, "test.log", test.NewOffset(0), []byte(`{"key1":"value_1"}`))
 
 	wg.Wait()
 	p.Stop()