--- a/plugin/input/dmesg/dmesg.go
+++ b/plugin/input/dmesg/dmesg.go
@@ -111,7 +111,7 @@ func (p *Plugin) read() {
 
 		out = root.Encode(out[:0])
 
-		p.controller.In(0, "dmesg", Offset(ts), out, false, nil)
+		p.controller.In(0, "dmesg", pipeline.NewOffsets(ts, nil), out, false, nil)
 	}
 }
 
@@ -133,14 +133,4 @@ func (p *Plugin) Commit(event *pipeline.Event) {
 // PassEvent decides pass or discard event.
 func (p *Plugin) PassEvent(event *pipeline.Event) bool {
 	return true
-}
-
-type Offset int64
-
-func (o Offset) Current() int64 {
-	return int64(o)
-}
-
-func (o Offset) ByStream(_ string) int64 {
-	panic("unimplemented")
 }