--- a/plugin/input/http/http.go
+++ b/plugin/input/http/http.go
@@ -556,10 +556,10 @@ func (p *Plugin) processChunk(sourceID pipeline.SourceID, readBuff []byte, event
 
 		if len(eventBuff) != 0 {
 			eventBuff = append(eventBuff, readBuff[nlPos:pos]...)
-			_ = p.controller.In(sourceID, "http", Offset(pos), eventBuff, true, meta)
+			_ = p.controller.In(sourceID, "http", pipeline.NewOffsets(int64(pos), nil), eventBuff, true, meta)
 			eventBuff = eventBuff[:0]
 		} else {
-			_ = p.controller.In(sourceID, "http", Offset(pos), readBuff[nlPos:pos], true, meta)
+			_ = p.controller.In(sourceID, "http", pipeline.NewOffsets(int64(pos), nil), readBuff[nlPos:pos], true, meta)
 		}
 
 		pos++
@@ -568,7 +568,7 @@ func (p *Plugin) processChunk(sourceID pipeline.SourceID, readBuff []byte, event
 
 	if isLastChunk {
 		// flush buffers if we can't find the newline character
-		_ = p.controller.In(sourceID, "http", Offset(pos), append(eventBuff, readBuff[nlPos:]...), true, meta)
+		_ = p.controller.In(sourceID, "http", pipeline.NewOffsets(int64(pos), nil), append(eventBuff, readBuff[nlPos:]...), true, meta)
 		eventBuff = eventBuff[:0]
 	} else {
 		eventBuff = append(eventBuff, readBuff[nlPos:]...)
@@ -724,13 +724,3 @@ func stringToUUID(input string) (uuid.UUID, error) {
 
 **`request_uuid`**  *`string`*
 }*/
-
-type Offset int64
-
-func (o Offset) Current() int64 {
-	return int64(o)
-}
-
-func (o Offset) ByStream(_ string) int64 {
-	panic("unimplemented")
-}