… plugin/action/set_time/set_time_test.go
… plugin/action/split/split_test.go
… plugin/action/throttle/throttle_test.go
… plugin/input/file/offset_test.go
… plugin/input/file/slicemap.go
… plugin/input/file/worker_test.go
… plugin/input/k8s/k8s_test.go
… test/file_base.go
… test/test.go